+++ src/agent/onefuzz-agent/src/managed/mod.rs
@@ -1,0 +1,1 @@
+pub mod cmd;
