+++ src/agent/onefuzz-supervisor/src/worker.rs
@@ -215,7 +215,7 @@ impl IWorkerRunner for WorkerRunner {
 
         let mut cmd = Command::new("onefuzz-agent");
         cmd.current_dir(&working_dir);
-        cmd.arg("-c");
+        cmd.arg("managed");
         cmd.arg("config.json");
         cmd.stderr(Stdio::piped());
         cmd.stdout(Stdio::piped());
